feat: delete CPS interpreter, simplify coroutine state
Remove the entire CPS tree-walking interpreter (~4,400 lines) that was
made redundant by Phase 1's first-class continuations in the bytecode VM.

Deleted:
- src/compiler/cps/ (13 files): CpsExpr, CpsTransformer, CpsInterpreter,
  Trampoline, Action, Continuation, ContinuationArena, ContinuationPool,
  CpsJitCompiler, mixed_calls, jit_action

Simplified Coroutine struct (value_old/mod.rs):
- Removed saved_continuation (CPS Rc<Continuation>)
- Removed saved_env (CPS environment)
- Removed saved_context and CoroutineContext (old bytecode path)
- Migrated yielded_value from OldValue to new NaN-boxed Value

Simplified coroutine resume (primitives/coroutines.rs):
- Created state: always uses bytecode path (no CPS decision)
- Suspended state: always uses saved_value_continuation (no fallbacks)
- Removed execute_coroutine_cps, resume_coroutine_cps,
  register_global_effects
- Moved old_value_to_new/new_value_to_old to coroutines module
  (still needed for broader value migration)

Simplified VM (vm/mod.rs, vm/core.rs):
- Yield handler no longer creates CoroutineContext
- Removed resume_from_context method

Removed CPS references from Cranelift JIT (jit_compile.rs)
