+++ proptest-regressions/property/coroutines.txt
@@ -1,0 +1,7 @@
+# Seeds for failure cases proptest has generated in the past. It is
+# automatically read and these particular cases re-run before any
+# novel cases are generated.
+#
+# It is recommended to check this file in to source control so that
+# everyone who runs the test benefits from these saved cases.
+cc 41f490ebdc106bb0590f99e66c5b13f20a51dc04a959be52c107b0fd75f5127d # shrinks to num_yields = 1
