--- src/compiler/cps/jit.rs
@@ -1,201 +1,0 @@
-//! JIT compilation for CPS expressions
-//!
-//! This module compiles CPS-transformed expressions to native code
-//! that returns JitAction values.
-
-use super::CpsExpr;
-use crate::effects::Effect;
-
-/// CPS JIT compiler
-pub struct CpsJitCompiler;
-
-impl CpsJitCompiler {
-    /// Check if a closure should use CPS compilation
-    pub fn should_use_cps(effect: Effect) -> bool {
-        effect.may_yield()
-    }
-
-    /// Check if a function can be inlined
-    pub fn can_inline(func: &CpsExpr) -> bool {
-        match func {
-            CpsExpr::GlobalVar(_sym) => {
-                // Check if it's a known primitive
-                // For now, we'll return false as we need symbol table access
-                // to check if a symbol is a primitive
-                false
-            }
-            _ => false,
-        }
-    }
-
-    /// Compile a CPS expression to native code
-    ///
-    /// This is a placeholder for the full implementation.
-    /// Returns a result indicating success or error.
-    pub fn compile_cps_expr(expr: &CpsExpr) -> Result<(), String> {
-        match expr {
-            CpsExpr::Literal(_value) => {
-                // Literal values become Done actions
-                Ok(())
-            }
-
-            CpsExpr::Var { .. } => {
-                // Load variable, create Done action
-                Ok(())
-            }
-
-            CpsExpr::GlobalVar(_sym) => {
-                // Load global, create Done action
-                Ok(())
-            }
-
-            CpsExpr::Pure {
-                expr: _expr,
-                continuation: _continuation,
-            } => {
-                // Compile pure expression, then apply continuation
-                Ok(())
-            }
-
-            CpsExpr::Yield {
-                value: _value,
-                continuation: _continuation,
-            } => {
-                // Compile value, create Yield action
-                Ok(())
-            }
-
-            CpsExpr::PureCall {
-                func: _func,
-                args: _args,
-                continuation: _continuation,
-            } => {
-                // Compile pure call, apply continuation
-                Ok(())
-            }
-
-            CpsExpr::CpsCall {
-                func: _func,
-                args: _args,
-                continuation: _continuation,
-            } => {
-                // Create Call action for CPS call
-                Ok(())
-            }
-
-            CpsExpr::Let {
-                index: _index,
-                init: _init,
-                body: _body,
-            } => {
-                // Compile let binding
-                Ok(())
-            }
-
-            CpsExpr::If {
-                cond: _cond,
-                then_branch: _then_branch,
-                else_branch: _else_branch,
-                continuation: _continuation,
-            } => {
-                // Compile conditional
-                Ok(())
-            }
-
-            CpsExpr::Sequence {
-                exprs: _exprs,
-                continuation: _continuation,
-            } => {
-                // Compile sequence
-                Ok(())
-            }
-
-            CpsExpr::While {
-                cond: _cond,
-                body: _body,
-                continuation: _continuation,
-            } => {
-                // Compile while loop
-                Ok(())
-            }
-
-            CpsExpr::For {
-                index: _index,
-                iter: _iter,
-                body: _body,
-                continuation: _continuation,
-            } => {
-                // Compile for loop
-                Ok(())
-            }
-
-            CpsExpr::And {
-                exprs: _exprs,
-                continuation: _continuation,
-            } => {
-                // Compile and expression
-                Ok(())
-            }
-
-            CpsExpr::Or {
-                exprs: _exprs,
-                continuation: _continuation,
-            } => {
-                // Compile or expression
-                Ok(())
-            }
-
-            CpsExpr::Cond {
-                clauses: _clauses,
-                else_body: _else_body,
-                continuation: _continuation,
-            } => {
-                // Compile cond expression
-                Ok(())
-            }
-
-            CpsExpr::Lambda {
-                params: _params,
-                body: _body,
-                captures: _captures,
-                num_locals: _num_locals,
-            } => {
-                // Compile lambda
-                Ok(())
-            }
-
-            CpsExpr::Return(_value) => {
-                // Return action
-                Ok(())
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::value::SymbolId;
-
-    #[test]
-    fn test_should_use_cps() {
-        assert!(!CpsJitCompiler::should_use_cps(Effect::Pure));
-        assert!(CpsJitCompiler::should_use_cps(Effect::Yields));
-        assert!(!CpsJitCompiler::should_use_cps(Effect::Polymorphic(0)));
-    }
-
-    #[test]
-    fn test_can_inline_global_var() {
-        let expr = CpsExpr::GlobalVar(SymbolId(1));
-        // For now, can_inline returns false for all globals
-        // This will be enhanced when we have symbol table access
-        assert!(!CpsJitCompiler::can_inline(&expr));
-    }
-
-    #[test]
-    fn test_can_inline_literal() {
-        use crate::value::Value;
-        let expr = CpsExpr::Literal(Value::int(42));
-        assert!(!CpsJitCompiler::can_inline(&expr));
-    }
-}
