+++ src/compiler/mod.rs
@@ -5,7 +5,6 @@ pub mod bytecode_debug;
 pub mod capture_resolution;
 pub mod compile;
 pub mod converters;
-pub mod cps;
 pub mod cranelift;
 pub mod jit_coordinator;
 pub mod jit_executor;
@@ -19,7 +18,6 @@ pub mod symbol_index;
 pub use bytecode::{Bytecode, Instruction};
 pub use bytecode_debug::{disassemble, format_bytecode_with_constants};
 pub use converters::value_to_expr;
-pub use cps::{Action, Continuation};
 pub use jit_coordinator::JitCoordinator;
 pub use jit_executor::JitExecutor;
 pub use jit_wrapper::{compile_jit, is_jit_compilable, JitCompiledFunction};
